py/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==0.15.7             # via -r requirements.in
avro-python3==1.9.2.1     # via -r requirements.in
azure-core==1.6.0         # via azure-cosmos, azure-eventhub, azure-storage-blob, azure-storage-file-datalake
azure-cosmos==4.0.0       # via -r requirements.in
azure-eventhub==5.1.0     # via -r requirements.in
azure-storage-blob==12.3.2  # via -r requirements.in, azure-storage-file-datalake
azure-storage-file-datalake==12.0.2  # via -r requirements.in
certifi==2020.6.20        # via msrest, requests, uamqp
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
confluent-kafka==1.4.2    # via -r requirements.in
cryptography==2.9.2       # via azure-storage-blob
docopt==0.6.2             # via -r requirements.in
fastavro==0.23.5          # via -r requirements.in
idna==2.9                 # via requests
isodate==0.6.0            # via msrest
msrest==0.6.16            # via azure-storage-blob, azure-storage-file-datalake
numpy==1.19.0             # via pandas, pyarrow
oauthlib==3.1.0           # via requests-oauthlib
pandas==1.0.5             # via -r requirements.in
pyarrow==0.17.1           # via -r requirements.in
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via arrow, pandas
pytz==2020.1              # via fastavro, pandas
requests-oauthlib==1.3.0  # via msrest
requests==2.24.0          # via -r requirements.in, azure-core, msrest, requests-oauthlib
six==1.15.0               # via azure-core, azure-cosmos, cryptography, isodate, python-dateutil, uamqp
uamqp==1.2.8              # via azure-eventhub
urllib3==1.25.9           # via requests
